@@ -13,17 +13,17 @@
   "license": "ISC",
   "dependencies": {
     "jsstore": "^2.0.5",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
+    "@babel/core": "^7.2.2",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/preset-react": "^7.0.0",
+    "babel-loader": "^8.0.5",
+    "file-loader": "^3.0.1",
     "webpack": "^4.6.0",
-    "webpack-cli": "^2.0.15",
-    "webpack-dev-server": "^3.1.3",
-    "worker-loader": "^1.1.1"
+    "webpack-cli": "^2.0.15"
   }
 }
